finished auth, added to azure cloud

+++ Controllers/AuthorsController.cs
@@ -1,22 +1,27 @@
-﻿using bookstore.Data.Dtos.Authors;
+﻿using bookstore.Auth.Model;
+using bookstore.Data.Dtos.Authors;
 using bookstore.Data.Entities;
 using bookstore.Data.Repositories;
+using Microsoft.AspNetCore.Authorization;
 using Microsoft.AspNetCore.Mvc;
+using Microsoft.IdentityModel.JsonWebTokens;
+using System.Security.Claims;
 
 namespace bookstore.Controllers
 {
-
     [ApiController]
     [Route("publishers/{publisherId}/authors")]
     public class AuthorsController : ControllerBase
     {
         private readonly IPublishersRespository publishersRespository;
         private readonly IAuthorsRepository authorsRepository;
+        private readonly IAuthorizationService authorizationService;
 
-        public AuthorsController(IPublishersRespository publishersRespository, IAuthorsRepository authorsRepository)
+        public AuthorsController(IPublishersRespository publishersRespository, IAuthorsRepository authorsRepository, IAuthorizationService authorizationService)
         {
             this.publishersRespository = publishersRespository;
             this.authorsRepository = authorsRepository;
+            this.authorizationService = authorizationService;
         }
 
         [HttpGet]
@@ -39,6 +44,7 @@ namespace bookstore.Controllers
         }
 
         [HttpPost]
+        [Authorize(Roles = BookstoreRoles.User)]
         public async Task<ActionResult<AuthorDto>> Create(CreateAuthorDto createAuthorDto, int publisherId)
         {
             var publisher = await publishersRespository.GetAsync(publisherId);
@@ -52,7 +58,8 @@ namespace bookstore.Controllers
                 LastName = createAuthorDto.LastName,
                 BornDate = createAuthorDto.BornDate,
                 Nationality = createAuthorDto.Nationality,
-                Publisher = publisher
+                Publisher = publisher,
+                UserId = User.FindFirstValue(JwtRegisteredClaimNames.Sub)
             };
 
             await authorsRepository.CreateAsync(author);
@@ -61,13 +68,22 @@ namespace bookstore.Controllers
                 author.FirstName, author.LastName, author.BornDate, author.Nationality));
         }
 
-        [HttpPut("{authorId}")]
+        [HttpPut]
+        [Route("{authorId}")]
+        [Authorize(Roles = BookstoreRoles.User)]
         public async Task<ActionResult<AuthorDto>> Update(UpdateAuthorDto updateAuthorDto, int authorId, int publisherId)
         {
             var author = await authorsRepository.GetAsync(publisherId, authorId);
 
             if (author == null)
                 return NotFound();
+
+            var authorizationResult = await authorizationService.AuthorizeAsync(User, author, PolicyNames.ResourceOwner);
+
+            if (!authorizationResult.Succeeded)
+            {
+                return Forbid();
+            }
 
             author.FirstName = updateAuthorDto.FirstName;
             author.LastName = updateAuthorDto.LastName;
@@ -77,13 +93,22 @@ namespace bookstore.Controllers
             return Ok(new AuthorDto(author.Id, author.FirstName, author.LastName, author.BornDate, author.Nationality));
         }
 
-        [HttpDelete("{authorId}")]
+        [HttpDelete]
+        [Route("{authorId}")]
+        [Authorize(Roles = BookstoreRoles.User)]
         public async Task<ActionResult> Delete(int publisherId, int authorId)
         {
             var author = await authorsRepository.GetAsync(publisherId, authorId);
 
             if (author == null)
                 return NotFound();
+
+            var authorizationResult = await authorizationService.AuthorizeAsync(User, author, PolicyNames.ResourceOwner);
+
+            if (!authorizationResult.Succeeded)
+            {
+                return Forbid();
+            }
 
             await authorsRepository.DeleteAsync(author);
